Sftp and Parallel Sftp actions will skip transfer if the file already is present on the target
Before beginning transfer, these actions will ask the target machine to compute the SHA1 hash of the file on the target side. If the SHA1 matches, then both size and content of the file on the target side is as desired - and therefore there is no reason to transfer the file.
This additional logic makes it easy to re-run the transfer operation in case only a few of the targets failed during transmission.