SendorJob has been eliminated. Now there are only SendorTasks and SendorActions.
The job concept was initially intended to group a set of tasks, enable the user to manage a group of tasks together in the UI, and manage dependencies between tasks. The application has not taken advantage of any of those features so far -- and therefore the job will be eliminated.
A positive effect of eliminating jobs is that multiple distribution requests (done at nearly the same time, by separate people) will now run in parallel.

Cancelling tasks is currently not possible. Needs some extra work to enable that.