Introduced Backbone.js & Underscore.js. Tasks are now rendered clientside, and refreshed every 10 seconds.
Backbone's Model/View architecture is used for handling tasks on the Javascript side.
The models are quite straightforward.

The view for the collection can only rebuild itself entirely or add individual items; removal of items is not yet implemented.

The template for the individual task is currently stored inside of index.html as I haven't figured out a good way to move it to its own file yet.

The clientside model is refereshed every 10 seconds through a fetch operation. This fetch will download the full set of tasks and is thus a rather expensive operation. A differential feed of some kind (for example: a websocket feed of any changes) would present a more scalable solution.