Implemented cleanup job for tasks
Tasks that have been waiting to get processed for more than a certain time will automatically be cancelled.
Tasks that are older than a certain time will be removed from the list altogether.

Also, renamed the SendorTask.set_queue_info operation to SendorTask.enqueued to better reflect its purpose (as it now logs the timestamp of the enqueued action).