Implemented a toggle for showing/hiding the log section of a task
The show/hide status is currently stored inside of the model. This is technically not the right place for the information, but if the value is to be kept within the view then the collection needs to be rewritten to not destroy & recreate all views when a single element is added/removed.